Suppress bare repositories in RepositorySearchDialog
When a bare repository is selected in the GitSelectRepositoryPage, the
import wizard throws and reports multiple NoWorkTreeExceptions.

Since a bare repository has no working tree, one cannot import from a
bare repository. Therefore, suppress bare repositories in the
RepositorySearchDialog which is used to add repositories to the
GitSelectRepositoryPage when used from the GitImportWizard. The
user thus cannot select a bare repository.

Bug: 495930
Change-Id: I6102eadd19e07ae47cf48ce0bf24ef6e64a44749
Signed-off-by: Matthias Sohn <matthias.sohn@sap.com>