Update patch set 2
Patch Set 2:

> I like the new widget, nice :-)
 > 
 > could you
 > * adapt styling to the underlying form-based view which has a
 > pretty whitish background
 > * suppress the shadow
 > see http://imgur.com/qlcsIaz
 > 
 > I think the widget should recompute the search result when the
 > selection in the combo changes, e.g. from "comments" to "author".
 > Also when the widget has a search term and then is hidden and shown
 > again the search result should be recomputed automatically

"background": will try, but I suspect this will be quite a bit of work (theme changes)... hard-coding it to "whitish" won't do.

"shadow": if there is a style constant in SWT for this, yes. Otherwise, no. I'd like to stay away from platform-specific stuff. The shadow is part of the SWT shell implementation on Mac. On GTK, there is no such shadow (see Lars's screenshot).

"recomputing": (1) on search scope drop-down changes: OK. Will have to check whether the old implementation did so. (2) on re-open: not sure if that's really convenient. It can be done, of course.

Patch-set: 2