Keyboard shortcuts for the rebase interactive view
As suggested in the enhancement request:
* 'e' -- edit
* 'f' -- fixup
* 'p' -- pick
* 'r' -- reword
* 's' -- squash
* DEL -- skip
* Ctrl/Cmd-Arrow_Up and Ctrl/Cmd-Arrow_Down for moving lines up/down.

Shortcuts are displayed in the context menu and in the tooltips of the
toolbar items.

These actions are a bit special since they react on unmodified
keystrokes. At least on Mac and on GTK there is some navigation
behavior on plain keystrokes: when a key is pressed, the selection
jumps to the next row that has a cell that starts with that key.
On GTK, there's even a text input pop-up that steals the focus in
this case. Since our menu also reacts on unmodified keys, it is
rather confusing if the selection changes when one hits a key not
handled here. Disable this, but allow normal table/tree navigation
keys to pass through. The "prefix navigation" is still available
by typing upper-case letters (shift + key).

Bug: 460595
Change-Id: I21846b8b5b05262d66dcb388c592b5e4cd24a8fa
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>