Update patch set 1
Patch Set 1:

> Not sure if that should be done differently but I noticed that I
 > loose the focus if I press an unassigned key, e.g., a.

Cannot reproduce on Mac. I have noticed that there is some kind of search going on on unhandled keys, which selects some other rows, but so far I've been unable to recognize what the rules are :-), and I haven't looked yet where that behavior comes from.

On GTK, this search opens a text input box in the lower right-hand corner of the rebase interactive view, and focuses that box, which moreover disappears on its own if there's no further input. When it disappears, it doesn't re-set the focus to where it was before.

Guess we'll have to find a way to disable this search. It's not very helpful in the rebase interactive view.

Patch-set: 1
Label: Verified=0