Update patch set 4
Patch Set 4: Code-Review-1

> Could someone test this on Windows, please? That's the one platform
 > that I don't have...

On windows 10 it looks not nice. The find box hovers over the view content descripton always in the left upper corner and has a bigger height so overlapping it. Initially it has some kind of shadowed borders, but after resizing etc this shadow goes away.  If I will get time I will post a screenshot, but as it is today this looks weird (not because of shaddows but because of the strange position).

Patch-set: 4
Reviewer: Gerrit User 1179 <1179@97ee7c02-f12f-4043-b43e-dea463d88b31>
Label: Code-Review=-1