Update patch set 5
Patch Set 5:

I really like Matthias's idea of using the content assist to cache the user's message. That way we don't need to decide whether to overwrite the message, we just do it on every task activation (which provides an easy way to regenerate) and the user can always get back what was there before if they want to.

Patch-set: 5
Label: Code-Review=0