Update patch set 1
Patch Set 1: -Code-Review

> did you also try your favourite rebase interactive use cases

Not in the original test. Now I tried it and it does not work. If you commit should also support this in interactive rebase I would expect to see the committed files also in the "Staged Changes" view, if I edit one commit. But my "Staged Commit" is empty.

Patch-set: 1