Update patch set 2
Patch Set 2:

(1 comment)

> Thanks for the clarification. This can than be handled by another
 > bug.

Not sure it _is_ a bug. Testing on my Mac seems to indicate that this is some kind of accessibility or built-in keyboard navigation thing. It also happens in the project explorer, and also outside of Eclipse in the Mac's Finder. In a tree control on Mac, typing a character navigates to the next row that has a cell starting with that character. And if you type fast enough to the row (cell) starting with the typed character sequence.

I have now disabled this in the rebase interactive view. BTW, that focus thief (text pop-up) on GTK appears to be non-functional. One can enter text, but then nothing happens.

Patch-set: 2
Label: Verified=0