Update patch set 1
Patch Set 1:

> Doing this here -- hard-wired (even with a check whether it's available) -- just doesn't look right.

I hope that long term the EGit import wizard can be retired and only the platform import wizard is used. The need for custom import wizard should be removed with our new SmartImport Wizard. Not sure if it is worth to design an extensible solution for that.

Patch-set: 1