Test stability: eliminate one arbitrarily timed wait
testPushToOrigin() sometimes failed with a WidgetDisposed exception.
Apparently there is some refresh of the viewer occurring while the test
is trying to expand a node hierarchy.

It's not entirely clear why this happens here, nor why the test was
waiting for jobs for at most 5 seconds. Let's try instead opening the
view before renaming the branch. Renaming the branch then should indeed
trigger a viewer refresh, which we can explicitly wait for.

Change-Id: I5c0ee9f3be81d8ed06be13747d5d888cfa4b1037
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>