Fix expansion problems and provide auto selection support.
The changes also include enhancements. When you have a large set of
changes, but you only want to selectively commit some of them, each time
you drag from unstaged to staged, both views scroll back to the top, and
you lose where you were reviewing the changes.

Enhancements include ensuring that collapse all is always available,
except for the list view as well as decoupling the expand all item
visibility so that it's only unavailable in the specific view with too
many items.

Another enhancement is that when you drag something from one view to
another, the result will be selected (and revealed) in the target view.
The selection is smart to select only a parent if all the children were
newly added, i.e., the selection designed to be minimal with the design
intent being that if selected added things are immediately dragged back
again, the operation will be undone.  The logic also sets a new
selection in source the view that otherwise no longer has a selection.  

Yet another enhancement is that when the viewer input is changed, there
is logic track the top item in the tree so that it can be restored after
the viewer update.  Then ensures that the scroll position is as stable
as possible and one can stage and unstage things (or refresh the
repository state) without losing your context.

One very bad behavior that's also fixed is that the
TreeViewer.getAllExpandedElements returns elements that aren't visible,
so updates often result in the tree becoming fully (way too) expanded
after each update. 

Also, switching presentations resulted in different expand-all and
collapse-all states than you'd get after an update.  One path considered
the total items in both views, removing the actions in both views, and
the other path considered whether either of the two views has too many
items and then removed both actions for both views.  I believe the
current implementation is much more consistent.  

Another bad behavior that's also fixed is that the recorded additional
elements to expand (computing while staging or unstaging the selection)
were not reset after they were used.  This also resulted in elements
that just kept expanding for no apparent good reason.

Bug: 451216
Change-Id: I748f61ec41fe34b3cee9612ed8dc156bd29b5812
Signed-off-by: Ed Merks <ed.merks@gmail.com>