Update patch set 2
Patch Set 2:

> > Build Started https://hudson.eclipse.org/egit/job/egit.gerrit/8951/
 > 
 > Sorry, the test seem to be unrelated to your changes. I've
 > retriggered hudson build. Unfortunately our UI tests are not as
 > stable as they should be :-(

Yes, the first build was good.  In any case, I tried to run it locally, and it passed.

BTW, I'll also update the Oomph setup so it works for master again.  I needed to include some more swtbot things to make the test run, so that was good to discover.

Patch-set: 2