Provide a remote mapping context to ModelProviders
Allow model providers to use remote information when trying
to determine whether a file is a part of a logical model or
not. When a file has been deleted locally, or when a file
has been created on one of the remote sides, the local side
is not enough to compute a full model. ModelProviders will
need access to the remote file data to take these locally
unknown files into account.
An interface to allow ModelProviders to be ignored in synchronizations
has been added.

Change-Id: I956dfb13093781accca36142cd87cf01b479fe1a
Also-by: Laurent Goubet <laurent.goubet@obeo.fr>
Also-by: Axel Richard <axel.richard@obeo.fr>
Also-by: Alexandra Buzila <abuzila@eclipsesource.com>
Signed-off-by: Laurent Delaigue <laurent.delaigue@obeo.fr>