Update patch set 2
Patch Set 2:

(1 comment)

I've addressed all the issues you've raised.  I also liked your idea of selecting a subsequent sibling instead of the parent in the case of removing the first child of some parent.

With regard to the indexDiff testing, I noticed that if you have an editor open for some unstaged resource, there are many calls to setInput, and the first few have indexDiff == null, so this test is needed to avoid selecting everything for initial population of the viewer.

I'm happy to make any further changes you think would be an improvement.

Patch-set: 2