Update patch set 4
Patch Set 4:

I added logic to record the overall cost of the setStagingViewerInput as well as the overall cost of just inner setInput and expandPreviousExpandedAndPaths, i.e., the later being the cost that existed before the change, and the former being that plus all the new logic to compute old and new content for view change analysis as well as tree traversal for scroll correction.  With a very large set of changes (CDO's repo with ~10,000 unstaged changed files and ~50 staged changes) I get results like this:

Full viewer update took 1744ms of which the setInput and reexpansion took 1689ms
Full viewer update took 27ms of which the setInput and reexpansion took 26ms

So the overhead is absolutely minimal even for this very large scenario.

Using the jgit repo with roughly 30 added/removed/changed items and dragging things back and forth produces results like this:

Full viewer update took 2ms of which the setInput and reexpansion took 1ms
Full viewer update took 1ms of which the setInput and reexpansion took 0ms
Full viewer update took 2ms of which the setInput and reexpansion took 2ms
Full viewer update took 0ms of which the setInput and reexpansion took 0ms
Full viewer update took 8ms of which the setInput and reexpansion took 6ms
Full viewer update took 1ms of which the setInput and reexpansion took 1ms
Full viewer update took 15ms of which the setInput and reexpansion took 10ms
Full viewer update took 1ms of which the setInput and reexpansion took 1ms
Full viewer update took 11ms of which the setInput and reexpansion took 9ms
Full viewer update took 1ms of which the setInput and reexpansion took 1ms
Full viewer update took 14ms of which the setInput and reexpansion took 12ms
Full viewer update took 0ms of which the setInput and reexpansion took 0ms
Full viewer update took 21ms of which the setInput and reexpansion took 19ms
Full viewer update took 1ms of which the setInput and reexpansion took 0ms
Full viewer update took 22ms of which the setInput and reexpansion took 16ms
Full viewer update took 2ms of which the setInput and reexpansion took 1ms
Full viewer update took 8ms of which the setInput and reexpansion took 7ms
Full viewer update took 20ms of which the setInput and reexpansion took 18ms
Full viewer update took 2ms of which the setInput and reexpansion took 2ms
Full viewer update took 25ms of which the setInput and reexpansion took 23ms
Full viewer update took 30ms of which the setInput and reexpansion took 23ms
Full viewer update took 9ms of which the setInput and reexpansion took 9ms
Full viewer update took 26ms of which the setInput and reexpansion took 21ms
Full viewer update took 2ms of which the setInput and reexpansion took 2ms
Full viewer update took 26ms of which the setInput and reexpansion took 24ms
Full viewer update took 1ms of which the setInput and reexpansion took 1ms
Full viewer update took 16ms of which the setInput and reexpansion took 15ms
Full viewer update took 4ms of which the setInput and reexpansion took 3ms

Here all the update costs, regardless of the overhead for the new logic, are so small that one could not possibly perceive a time difference.  In the end, a 5-10% overhead appears to be the case, i.e., in the worst case adding a 1/10 of a second to an operation that takes more than 1.5 seconds.

Patch-set: 4