Update patch set 1
Patch Set 1:

Thomas,

Have you installed an IDE using the EGit setup?  There are instructions here for the platform: 

https://wiki.eclipse.org/Platform_UI/How_to_Contribute

You'd do that same thing, except pick EGit's setup on the project's page.  Be sure to select the appropriate style of URI you want to access the repo, probably the ssh one, but who knows what you prefer.  This should result in a fresh new installation with the previous setup I contributed.  In the eclipse.ini there will be an entry like this:

-Doomph.redirection.egit=https://git.eclipse.org/c/egit/egit.git/plain/tools/oomph/EGit.setup->file:/D:/sandbox/USER-HOME/egit-master2/git/egit/tools/oomph/EGit.setup


This ensures that the setup used within that workspace is the one in the git clone itself. So if you have a environment setup up with Oomph, you can pull this change into your clone, and then do Help -> Perform Setup Tasks... to test it in that workspace.  Then you should see the working sets change to be the way Matthias designed them.

Patch-set: 1