Update patch set 8
Patch Set 8:

> @Thomas, are you interested in helping eclipse platform to design
 > such a widget for the search functionality? The current popup is
 > really annoying.

Interested yes, but, alas, I don't really have the time for it. Also, personally, I don't find the global search dialog that bad. A popup like the one prototyped here in this change has the basic problem that it obscures part of the view. That search could be improved is certain; otherwise there wouldn't be various plugins in marketplace that offer different ways of "insta-search". But hashing out all the nitty-gritty details of a general-purpose search widget is more than I could take on right now.

Patch-set: 8