Improve performance of StagingView a little
EGit creates a StagingEntry for each file that can appear in the staging
view. It creates all these objects up front, and each StagingEntry tries
immediately to determine the corresponding workspace IFile.

The latter is a rather expensive operation. When there are many
StagingEntries, this results in a potentially long delay before the
staging view populates, even if the tree or compact tree representation
is used and most of the files are in some not expanded folder and thus
not even visible.

By delaying finding the IFile until it's needed (which is when the file
actually is displayed in the staging view), the cost of this operation
is not incurred up front but only when nodes are expanded, and thus the
view populates much quicker initially.

If the list presentation is used, this change will not have any effect.

Bug: 500106
Change-Id: I6b7d4c7ce7be606bde6206255e710831721b15bb
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>