More efficient label provider for the staging view
The staging view's label provider was derived from
DecoratingLabelProvider, which is an ITreePathLabelProvider. For these,
the TreeViewer's WrappedViewerLabelProvider computes the tree path for
each element, which then ultimately isn't used because the
StagingViewLabelProvider is _not_ an ITreePathLabelProvider. This tree
path computation is, at least on GTK, fairly expensive.

By not deriving our label provider from DecoratingLabelProvider but
wrapping it, we can ensure that the framework does not compute these
tree paths that are then not used anyway, and decorations are
significantly faster on GTK.

Bug: 500106
Change-Id: If30bfbfeda6df2f89022995e8ba375be8c45ca3e
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>