Update patch set 1
Patch Set 1:

> nice performance improvement
 > 
 > Though I don't know why we need to differentiate between "Ignore"
 > and "Ignore folder". If the user selected a folder I would only
 > offer "Ignore Folder" but still name it "Ignore". Ignoring a folder
 > by adding individual entries for each files/folders contained in
 > the selected folder seems insane overhead.

Depends. Until now all operations in the staging view operated on files. People might have accustomed to that, and when they have a folder with, say, 3 unstaged files, actually might expect these three files to be ignored when the folder is selected, but not future additions in that folder.

To keep surprises for the user at a minimum, I opted for an extra "Ignore Folder" action instead.

Patch-set: 1
Label: Verified=0