Update patch set 3
Patch Set 3:

> I sent an email to Laurent asking if he can come up with a fix
 > quickly otherwise I think we should revert the complete change
 > until this regression is fixed

Thanks. 

Probably this will be better, but unfortunately I have no idea what the original patch is supposed to fix (commit message does not say me anything I can understand because I'm not familiar with ModelProviders) and how it is used by the rest of the world.

Patch-set: 3