Introduce SyncStrategy for logical model support
This allows to provide distinct implementations when doing
synchronizations depending on whether or not the user is interested in
logical models.
The default SyncStrategy implementation does not deal with logical
models at all and thus provides the best performance.
The ModelAwareSyncStrategy consults the logical model providers to
determine the set of resources involved in synchronization operations
and thus can have important performance impact especially on large
repositories.
Preference 'allow models to participate...' now defaults to false. This
has an impact on some test cases where we now set-up the preference
before running them.

Bug: 501990
Change-Id: I9322cef70c037e0b2c270a7c2bab7d2ffd278f60
Signed-off-by: Laurent Delaigue <laurent.delaigue@obeo.fr>