Introduce GitSynchronizer for logical model support
This allows providing distinct implementations when doing
synchronizations depending on whether or not the user is interested in
logical models.
The default synchronizer implementation does not deal with logical
models at all and thus provides the best performance.
The ModelAwareGitSynchronizer consults the logical model providers to
determine the set of resources involved in synchronization operations
and thus can have important performance impact especially on large
repositories. It uses a 'lazy' implementation of subscriber to load
additional resource revisions lazily when needed.

Bug: 501990
Change-Id: I9322cef70c037e0b2c270a7c2bab7d2ffd278f60
Signed-off-by: Laurent Delaigue <laurent.delaigue@obeo.fr>