Clean up "Show Blame" handling
* Replace deprecated ObjectContribution extension by Command/Handler
  combo and remove the default handler for the
  org.eclipse.egit.ui.team.ShowBlame command.
* Remove now unused legacy ShowBlameAction and the (wrong) command id
  in ActionCommands.

This does not yet remove the default handler for the history view's
org.eclipse.egit.ui.history.ShowBlame command. Apparently there were
once problems with creating this context menu declaratively (see bug
322751 and commit 44dfabcd); apparently having to do with the
availability of the selection. As a consequence this menu is built
programmatically but relying on some declarations (including default
handlers) in plugin.xml. These default handlers can still be invoked
outside of their intended context, but cannot be removed easily. The
approach taken for now for this ShowBlameHandler is thus to at least
prevent exceptions when it is invoked when it shouldn't be:

* Prevent NPE in the ShowBlameHandler (history view).

Bug: 495064
Change-Id: I3d570266cc58f873e512e124299b235d35ecfdb6
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>