Update patch set 1
Patch Set 1:

> This patch doesn't fix the problem.

That's very strange. I just tried with a fresh install of a Neon.1 Java package. I got the error as expected, then installed a local EGit build with the patch, and afterwards was able to clone and push via HTTP with authentication. As expected, I got EGit's authentication dialog in the clone wizard, and then both the clone and the subsequent push (with credentials from the secure storage) worked fine. I ran the same procedure again in a remote-debug session just to confirm, and the EGit Authenticator was left intact by MPC as I would have expected.

In any case, since we have now requested a simrel respin to ship the fix in MPC via automatic updates, I would suggest to abandon this change.

Patch-set: 1