Update patch set 1
Patch Set 1:

I think this is off-topic, but just to be thorough: One thing I did notice was that when cloning the repository without having credentials in secure storage, I got the EGit credentials dialog twice in the course of the Clone wizard, even though I chose to store my credentials: once when first listing the remote branches, and then again during the fetch after completing the wizard.

But I verified that the same happened for me even in an Eclipse SDK installation without MPC present, so I guess it is an unrelated issue in EGit.

Patch-set: 1