Fix the enablement of "Compare with->Index with HEAD"
This menu command makes sense only for staged files, but it was enabled
wrongly also on unstaged files. Apparently that was done on purpose in
commit 6d975be6 to avoid a costly index diff calculation to find out
whether the file was staged.

But if we are able to display "staged" decorations speedily, we
certainly are also able to determine quickly whether a file is staged.
Make isEnabled() properly check again, using the same mechanisms we use
for decorations, which rely on the IndexDiffCache. Remove the extra
check when the command is run, including the dialog.

Change-Id: Ic6ec1e407f2ba1403ae52bc09b4f2546f4e221c8
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>