Update patch set 3
Patch Set 3:

When using command "Pull..." on a local branch which has no upstream configuration yet the Pull wizard pre-fills the field "Reference" with the name of the currently checked out local branch, if there is no upstream branch with that name the pull command fails with the following error:

"Remote origin did not advertise Ref for branch refs/heads/b. This Ref may not exist in the remote or may be hidden by permission settings."

I think it would be better if the Pull wizard would check if the reference matching the current local branch exists upstream, and if it doesn't exist it should not pre-fill the field "Reference" to avoid the above error.

Patch-set: 3