Clean up "Compare With" actions
Replace two more legacy <objectContribution>s by command/handler combos.
Remove now unused legacy action classes, and remove a duplicate and
confusing command ID org.eclipse.egit.ui.team.CompareWithRevision. In
EGit, we have CompareWithRef and CompareWithCommit, and action IDs
follow the class names.

Fix comment in ResourceStatePropertyTester: since commit 2349a197 it
doesn't enforce the "single repository" condition anymore to allow "Add
to Index"/"Remove from Index" to work for multiple repositories at once.

Since the "Compare with" handlers must also be enabled on StagingEntries
that do not adapt to IResource (such as staged files not in the
workspace) and the "resourcesSingleRepository" property thus cannot be
used, introduce a new "selectionSingleRepository" property test in the
SelectionPropertyTester and use that.

Visual change: the ordering of the compare actions on a WorkingSet is
now consistent with the ordering on IResources. Previously and
inexplicably, the order was different.

Bug: 495064
Change-Id: I35277df4e82e70b0d8f6bd149c4218acd3f293f7
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>