Hide local branch upstream warning for remote branch upstream
When using a local branch as upstream, the branch creation dialog shows
a warning which was not removed even when switching to a remote branch
as upstream.

This regression was introduced with the fix for bug 440097. That one
removed the line "setMessage(null)". This fix reintroduces that line,
but as "setMessage(initialMessageWhenThePageWasCreated)" instead. This
avoids the previously experienced wrong height calculation of the
dialog, I have tested that extensively.

Bug: 452212
Change-Id: I849c483b5a8a42a80191c4105b102fe4cd5377c4
Signed-off-by: Michael.Keppler <Michael.Keppler@gmx.de>