Update patch set 7
Patch Set 7:

I'm rather busy @work right now. This is a big non-trivial change that needs time to review, and I guess I'll have to set up a test eclipse with EMF compare and some EMF models to test this. I won't get around to do an in-depth review in the next two weeks.

Feel free to add other reviewers; Matthias Sohn would know what to do about the CQ, if anything; and Andrey Loskutov might perhaps have the time to verify that this does not cause problems on the huge repos he has. I'd also be interested in seeing some numbers on typical models about the benefit of the incremental (or lazy) subscriber.

If I have a few hours, I might give that strange control flow I had mentioned a shot.

Best would be if we could make this quick enough so that we could leave the preference unchanged and leave model providers enabled by default. Disabling it by default feels a bit like cheating :-) Of course a badly written model provider (or a very complex one) may always slow down things, but then the user can flip the preference if needed. (Or bug the provider of the model provider to improve it...)

Patch-set: 7
Label: Verified=0