Update patch set 8
Patch Set 8:

Hello Andrey,

Could you please test this change-set that attempts to reintroduce logical model support in EGit on a huge git repo?
You'd need to check the preference 'Allow models to participate in synchronizations' since it's now off by default.
I tested it by adding a Thread.sleep in WorkingTreeIterator.isEntryIgnored but this is really not a satisactory test, so I'd really appreciate your feedback here so that we know if, at least, this new patch-set avoids the performance issue.

Thanks in advance!

Patch-set: 8