Update patch set 10
Patch Set 10:

Hello guys, thank you for the recent comments and launching the CQ!
I pushed a new patch set that reverts the default value of the property to 'true', as before.
It looks like the build failed because of an infrastructure problem, could someone retrigger it if the builds are back to normal?

Thanks in advance.

Laurent

Patch-set: 10
Label: Verified=0