Add an outline page for DiffEditorPage
Introduce a MultiPageEditorContentOutlinePage for the CommitEditor
that manages outline pages for nested editors. Give the nested
DiffEditorPage its own fully functional outline page. The outline
page is linked to the editor page: its selection updates when the
caret's position in the editor moves between file diffs. It also
has a simple context menu giving access to actions for opening the
files in an editor, or for showing a file diff in a compare editor.

Change-Id: Iddd72105b86c3a8c5d31e6fa874bb470ad1f964c