Update patch set 8
Patch Set 8:

(1 comment)

Nice. In the future I'd like to see this directory traversal be done completely in the background, populating the viewer and updating the messages as it goes, with the dialog controls active. Since that is going to be a somewhat larger refactoring (especially regarding stopping that background job) and will probably also need some changes in tests, I'll open a feature request for this.

Patch-set: 8
Label: Verified=0