user can change to local branch when fetching Gerrit changes
When attempting to fetch a Gerrit patch set which the user has already
fetched they are now presented with an option to change to the already
fetched branch. This is the default option if a local branch exists. If
no local branch exists for a ref this option is not available.

Bug: 509181
Change-Id: I81af2906fe44dacb35ca9c23e4cd2583750f3b8a
Signed-off-by: Jaxsun McCarthy Huggan <jaxsun.mccarthy@tasktop.com>