Fix focus restoration in CommitEditor pages
FormPage by default re-sets the focus to the first child of the
managed form. In the CommitEditorPage, this is not convenient;
it'll always set focus on the author indication. Add focus tracking
to identify the control currently having the focus so that it can
be re-focused when the page becomes active again after a page
change or editor de-activation.

Bug: 509486
Change-Id: I678bdfde9105d04669a42ad81502eb1064c470bc
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>
Also-by: Thomas Wolf <thomas.wolf@paranor.ch>
Signed-off-by: Markus Keller <markus_keller@ch.ibm.com>