Update patch set 2
Patch Set 2:

(1 comment)

To completely fix this, I guess one needs to move that focus tracking and custom setFocus() into a custom ScrolledComposite set as the only child of the ManagedForm. Then we can leave the shared header to the standard implementation (which, however, will always set the focus on _its_ first child: the commit id), and deal with our own Controls locally.

Patch-set: 2