Update patch set 3
Patch Set 3:

> Thomas, I see you introduced a merge commit for this change. Most
 > likely an error but I think you need to configure your branch to
 > use "rebase" during pull

Not an error, just a minor oversight of my part. My pull config is set to rebase already. To edit an existing Gerrit change, I don't pull. I do a "Fetch from Gerrit", which is _not_ a pull but a fetch & checkout of FETCH_HEAD. Would be quite annoying if it automatically rebased, actually. I just forgot to rebase manually before pushing to Gerrit again. And since the Gerrit project is configured as "merge when necessary", we got that merge commit.

At work, we've been using for years only projects configured as "rebase if necessary", and then Gerrit takes care of that. Works very well.

Patch-set: 3