Another batch of defaultHandlers removed
Add dedicated handlers with activeWhen expressions, set command
visibility to show only when enabled, and remove default handlers
for some more commands.

* Import Projects... (repositories view)
* Import Change Projects... (history view)
* Create Branch... (repositories view, on tags and refs)
* New Remote... (repositories view)
* Open In Editor (repositories view, on file nodes)
* Open (repositories view)
* Link With Selection (repositories view)

"Open" is a multi-purpose command that is also invoked by an
OpenListener on the repositories view. It is visible as "Open In Commit
Viewer" on stashed commits, and invoked by the open listener for files
and for stashed commits. Additionally, it is invoked by a
DoubleClickListener on tags and refs, where it does a checkout.

Bug: 495064
Change-Id: I71e7107b482d98968da443d5b5e270b05586b807
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>