Fix context menu on old/new line number ruler in DiffEditorPage
The ruler context menu would not appear on the line number ruler in
the DiffEditorPage. The menu is set on the outer CompositeRuler
containing all left columns, but none is set on our own nested
CompositeRuler, and a CompositeRuler's control only forwards typed
event listeners but not SWT-level event listeners.

Work around this by adding our own SWT.MenuDetect listener to the
nested columns that checks the correct control's menu.

Change-Id: I63fcb448d7978eaa44ca2fd3862b4cb1a2c144b4
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>