Update patch set 13
Patch Set 13:

(1 comment)

Works well if the history view is visible, but if not, this doesn't work. GenericHistoryView then sets the input via showLastSelectedElement(), which again is only a single resource. Also, the selection obtained inside GitHistoryPage is then that page's own selection. To reproduce:

* Have both history and git staging view in the bottom right part stack.
* Have the history view visible.
* Select multiple resources in package explorer -> history view properly shows combined history.
* Click git staging tab.
* Select some different resources in package explorer.
* Click history tab -> history view shows only the history of the first selected element.

Perhaps the GitHistoryPage would have to track the last selection itself.

Patch-set: 13
Label: Verified=0