Update patch set 2
Patch Set 2: Code-Review-1

(2 comments)

Double-click clears a multi-selection and reduces it to just the element on which the double-click occurred. This is confusing.

On GTK, if there are less repositories than lines in the viewer, the behavior on double click in the empty viewer area is very confusing. When the dialog initially opens, there is no selection in the viewer: a double-click in the empty area toggles the checkbox of the first element. Once there is a selection,double-clicking in the empty area toggles the most recently selected element, but lets the multi-selection stand. A double-click on an existing element reduces the selection to just that element and toggles only that element's checkbox.

On Mac, double-clicks in the empty area don't do anything.

Furthermore, double clicks do not update the "Finish" button state.

Finally, if such double-click behavior was desired and could be implemented to work better with multi-selections: shouldn't all this be done in the tree viewer implementation so that other instances of CheckboxTreeViewer could also benefit and the UI experience would remain uniform? I notice that CheckboxTreeViewer.handleDoubleSelect() has code that should already do this (for single selections, at least), but it doesn't appear to work.

Patch-set: 2
Reviewer: Gerrit User 44628 <44628@97ee7c02-f12f-4043-b43e-dea463d88b31>
Label: Code-Review=-1