Remove default handlers for checkout commands
Provide dedicated handlers with appropriate activeWhen clauses
instead.

The remaining default handler for org.eclipse.egit.ui.commit.Checkout
cannot be removed (yet). The command is used for the checkout button
in the CommitEditor's toolbar. This is in the header of a
SharedHeaderFormEditor, and the usual technique for eliminating
default handlers doesn't work in that case: the button always
remains disabled. Something strange with the site or the selection
provider is going on, or these toolbar buttons need some more
hook-up, but I don't see what's missing.

Bug: 495064
Bug: 511635
Change-Id: Ied73a4f921c57d7b6b1f5f3b6cf09812d10a50ba
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>