Remove the default handler for the checkout command in CommitEditor
Provide a dedicated handler with an appropriate activeWhen clause
instead. Somehow handler activation expressions are not properly
re-evaluated for these CommitEditor commands; seems to have
something to do with them being used in the header of a
SharedHeaderFormEditor. Add a PartListener to the CommitEditor that
forces an update of the toolbar on activations.

Bug: 495064
Bug: 511635
Change-Id: I8086b4ec328fb6b4eaa63a6f3b68ca0b075c5866
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>