Update patch set 2
Patch Set 2:

(32 comments)

Thanks to you both for taking time for this long review.
Don't be surprised if I left a few comments unanswered. I would like to get your opinion first on a few structural issues. I.e. how to split up long module classes, how to separate those changes into multiple commits, should the EASE integration be part of the core plug-in or rather in a separate feature (see Christians last mail on the list), and should we support as many operations as possible in UI as well as in headless mode, or should we only put operations in the GitUiModule if it is strictly necessary?

Patch-set: 2
Label: Verified=0