Update patch set 4
Patch Set 4:

A less intrusive approach would be to leave all the validation logic as it was. Don't try to auto-normalize while the user is typing. Give the dialogs a button (not a checkbox) "Fix branch name", which would be enabled (or perhaps even shown only) when the validator detects an error. When the user clicks that button, then normalize the input once (and re-validate).

No need for fiddling with text caret positions; no need for strange attempts to detect a paste, the user isn't surprised when consecutive blanks or underscores are collapsed to one, and no need to have multiple ModifyListeners (one of them even modifying the text again).

Perhaps add a VerifyListener that changes blanks to underscores.

Patch-set: 4