Update patch set 4
Patch Set 4:

> You are right. I should have worked with the VerifyListener all
 > along.

Actually, I'm not sure that would work well either. In any case it shouldn't call the normalizer: since that collapses consecutive _or - into one, we'd risk screwing up with the text cursor position whatever we do.

You could use a VerifyListener to prevent certain invalid characters, but to allow typing consecutively f-o-o-/-b-a-r, you'll even have to allow a / or a . at the end.

So the existing validator (modify listener) must remain anyway, and thus a VerifyListener is probably not worth the trouble and we'd be better off with a "Fix" button.

Or perhaps use content assist to propose a fixed name instead of a button, but this is likely to give a confusing UI when the branch name already is valid: user types a valid name, sees the content assist lightbulb, types Ctrl-Space, and then wonders why the popup shows the same name... unless we can enable content assist only when the name is invalid, and disable it (and hide the lightbulb) otherwise.

Patch-set: 4