Branch name normalization: use content assist
Rewrite the new branch name normalization to not interfere with the
user's typing. Instead just replace blanks by underscores. Use the
content assist framework to suggest an auto-corrected branch name if
the current input does not form a valid branch name. Show content
assist availability via the normal lightbulb decorator.

This content-assist can be triggered via the normal keystroke. On
characters that are known to be rejected by JGit in any case, it is
also triggered automatically. If the current input _does_ form a
valid branch name, content assist is disabled and the lightbulb is
hidden.

This gives the user full control over the input, but provides easily
accessible auto-correction suggestions.

Change-Id: I7d5b3a422823b58c6f91ff13898cfb1fb154b015
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>