Update patch set 14
Patch Set 14: Code-Review+1

Now it works nicely. I notice that we have different behavior in the history and in the staging view. The staging view doesn't react on multi-selections, and resets to "No repository selected" when the selection changes to an empty selection or one that doesn't translate to a repository. The GitHistoryPage does react to multi-selections and keeps whatever it currently displays when the selection becomes empty or doesn't translate to any repository.

That should be fixed in a follow-up change. Probably extract the selection tracking from the GitHistoryPage into a separate re-usable (and possibly customizeable) SelectionTracker that can then also be used in the StagingView. The implementation in GitHistoryPage is much simpler than the (old) part tracking in the StagingView.

I think this is OK to go in. Minor nit-pick: why are all the new methods in GitHistoryPage protected? They're not used anywhere else and could just as well be private.

@Matthias, could you please double-check?

Patch-set: 14
Label: Code-Review=+1