Update patch set 14
Patch Set 14:

Hi Thomas,

Thank you very much for your very fast review! 

You are right, the behavior between the staging view and history view is now slightly different, but should probably be adapted in a separate change. Do you want to keep it in the same bug or in a new bug specific to the staging view?

I also changed the method visibility to 'private' to be in line with the other methods. I tend to make helper methods, e.g., just getting the selection, protected to be more extensible. But you are right, in this case it is not necessary since no one will subclass this class and we do not use the methods anywhere else.

Thanks again!

Patch-set: 14
Label: Verified=0