Update patch set 16
Patch Set 16: Code-Review-1

Sorry, this doesn't work quite correctly yet. Without this change, the history view updates when an editor is activated and shows the history of the resource opened in that editor. With this change, there are cases where this no longer works.

To reproduce:

* Have at least two files in an EGit-managed project.
* Open one file (A) in an editor.
* Expand the project in the package explorer.
* Select both files (A and B, or even more files, but including A).
* History view shows the combined history.
* Activate the editor in which A is open. (Click on the tab, or in the text pane.)

Expected:

* History view updates to show the history of only A.

Actual:

* History view does not update and keeps showing the combined history of the selection in the package explorer.

The reason is twofold:

First, GitHistoryPage simply ignores any non-structured selections. An editor, however, reports an ITextSelection. Even if that is fixed (for instance by using SelectionUtils.getStructuredSelection()), it won't work because of:

Second, GenericHistoryView reacts on part activation and calls setInput with the correct object (resource A open in the editor). But part activation comes before the notification of the selection change. GitHistoryPage thus still has the previous selection from the package explorer as "lastSelection". mostFittingInput will then return that lastSelection as input because it includes A.

This is confusing: the user experience is that sometimes the history page updates and sometimes not. (It updates if the lastSelection does not contain A.)

Not sure how to fix this. Even if GitHistoryPage did its own part activation tracking (like in the StagingView), there's no guarantee about the order in which the two listeners (GitHistoryPage and GenericHistoryView) are notified.

There's also a second bug:

* Make sure the GitHistoryPage shows the history of a single file.
* Uncheck "Link with Editor and Selection".
* Select several files in the package explorer (doesn't matter whether or not that selection includes A).
* Select (Context menu)->Show In->History.

Expected:

* History view is activated and updates to show the combined history of the selected files.

Actual:

* History view is activated, does not update, and a beep occurs.

Patch-set: 16
Label: Code-Review=-1