Update patch set 16
Patch Set 16:

Hi Thomas,

You are right that in the cases you describe, the current implementation does not work, and thank you very much for also providing your findings for the reasons why. I also had a more detailled look at it and I see a few problems.

For the first bug, we may do the following (as included in the next patch set):
1. Include the active part in the selection tracking.
2. Clearing the selection once we have used it to avoid incorrect behavior when a part is activated. In this case, the fallback selection is used.

What worries me, is the second case that may not be possible to solve. The reason for this is that there are two ways the history page is updated:
1. Linked Selection: We track the active selection and update the view on demand. This is done by the GenericHistoryView which only extracts the first element of a selection on selection change and calls showLastSelectedElement on different occasions, e.g., when the selection changes or linking is enabled. This is the case our previous fix tried to solve by tracking our own selection and keeping all elements of the selection, not only the first.
2. Show In > History: In this case, the o.e.ui.internal.ShowInHandler is triggered which calls the show method of the GenericHistoryView to see if the view can show a certain context. The context contains the selection (e.g., the tree selection from the Project Explorer) and an input (e.g., the JavaModel). Again, the GenericHistoryView only considers the first element of a structured selection and otherwise tries to show the input. So when we select multiple elements, the selection is not considered and the context input does not supply a page source (like for the JavaModel), the history page is never updated. As opposed to the selection tracking, where we are called with the wrong/incomplete input, here we are not called at all. And there seems to be no way to influence/override the GenericHistoryView.show behavior.

This leads to the problem of seemingly inconsistent behavior from a user perspective. To be honest, I am not sure what is the best way to continue with this.

Patch-set: 16
Label: Verified=0