Update patch set 17
Patch Set 17:

(2 comments)

Works now. Some minor nitpicks.

However: did you notice that bug 392948 in EGit depends on bug 392949 in platform.team? In Team, there's also bug 406894 (which is the "Show In" problem), and bug 372959 (and its duplicate 388195). Plus there's another bug (not reported yet): the history view does not update when the selection changes in an MultiPageEditorPart. (For instance, switch between MANIFEST.MF and plugin.xml in a PDE editor: history will always show the history of plugin.xml.) That's because GenericHistoryView.editorActivated() doesn't account for multi-page editors. Even if we fixed that in the new GitSelectionTracker, it wouldn't work since we'd get MANIFEST.MF from it all right, but the GenericHistoryView will call GitHistoryPage.setInput() with plugin.xml, and since those don't overlap, we'll keep using plugin.xml.

Given that we

* have to build quite some infrastructure to track part activations and selections ourselves,
* run into the "Show In" problem (bug 406894), and
* should fix that MultiPageEditorPart problem,

how about fixing GenericHistoryView in platform.team instead of adding work-arounds in EGit?

The current change looks good and I like the factoring out of the tracking into a separate class. I'm OK with including it, given that we cannot deal with the "Show In" and the MultipageEditorPart problems locally. But really it's the GenericHistoryView that needs fixing. If and when it ever gets fixed, we might need to add a version check to do our own tracking only on platforms where we still have that broken GenericHistoryView.

Patch-set: 17