Update patch set 17
Patch Set 17:

(3 comments)

Hi Thomas,

Thanks again for you really fast feedback. You are right, I think the most promising solution would be to add support for multi-selection directly in Team. My only concern is that they seem to have chosen deliberately to restrict themselves to single selection and I am not sure why. So, I hope they would pass a change that introduces multi-selection. Thank you also for pointing me in the direction of the already existing bugs. Although most seem to have been around for a long time, they seem to pile up and it would definitely make sense to adapt the GenericHistoryView.

Regarding this change, I incorporated your minor fixes and also added support for the multi-page scenario in the selection tracker by implementing of the IPageChangeListener interface. Implementations of the IPartListener2 interface (like the selection tracker) are automatically also notified about page changes when they implement this interface. Currently, this selection is not used since we only retrieve the selection from the tracker when we are triggered by the GenericHistoryView and this is currently not happening for multi-page editors. We could react to this change ourselves in the GitHistoryPage by supporting listeners in the selection tracker, but maybe this is better for another change, if we need it. Nevertheless, I think it makes sense to include it in the tracker for now to track the selections as we would expect it. Depending on how the fix in the Team will look like, we should definitely think also about the version tracking you were mentioning. 

If you disagree that the implementation of the multi-page support should be in this change, I will remove it.

Patch-set: 17
Reviewer: Gerrit User 79641 <79641@97ee7c02-f12f-4043-b43e-dea463d88b31>
Label: Verified=0